Organize files into package
