+++ src/main/java/jimmy/AddDeadlineCommand.java
@@ -1,3 +1,5 @@
+package jimmy;
+
 public class AddDeadlineCommand extends Command {
     private final String description;
     private final String by;
