+++ src/main/java/jimmy/AddEventCommand.java
@@ -1,3 +1,4 @@
+package jimmy;
 public class AddEventCommand extends Command {
     private final String description;
     private final String from;
