+++ src/main/java/jimmy/AddTodoCommand.java
@@ -1,3 +1,4 @@
+package jimmy;
 public class AddTodoCommand extends Command {
     private final String description;
