+++ src/main/java/jimmy/Event.java
@@ -1,3 +1,4 @@
+package jimmy;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
