+++ src/main/java/jimmy/ExitCommand.java
@@ -1,3 +1,4 @@
+package jimmy;
 public class ExitCommand extends Command {
     public ExitCommand() {}
