+++ src/main/java/jimmy/JimmyException.java
@@ -1,3 +1,4 @@
+package jimmy;
 public class JimmyException extends Exception {
     public JimmyException(String message) {
         super(message);
