+++ src/main/java/jimmy/ListCommand.java
@@ -1,3 +1,4 @@
+package jimmy;
 public class ListCommand extends Command {
     public ListCommand() {}
