+++ src/main/java/jimmy/Storage.java
@@ -1,3 +1,4 @@
+package jimmy;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
