+++ src/main/java/jimmy/Task.java
@@ -1,3 +1,4 @@
+package jimmy;
 public abstract class Task {
     protected String name;
     protected boolean isCompleted;
