+++ src/main/java/jimmy/Todo.java
@@ -1,3 +1,4 @@
+package jimmy;
 public class Todo extends Task{
     public Todo(String name) {
         super(name);
